% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_variations.R
\name{calculate_variations}
\alias{calculate_variations}
\title{Calculate Variations}
\usage{
calculate_variations(base_value, variation_percent)
}
\arguments{
\item{base_value}{A numeric value representing the base.}

\item{variation_percent}{A numeric value representing the variation percentage.}
}
\value{
A list with `lower` and `upper` bounds.
}
\description{
Computes the lower and upper bounds for a given base value and variation percentage.
}
\author{
Sofia Daza
}
