% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone_radius_functions.R
\name{cone_radius}
\alias{cone_radius}
\title{Cone Radius Non-Vectorized Function}
\usage{
cone_radius(x)
}
\arguments{
\item{x}{A numeric value representing the height.}
}
\value{
A numeric value representing the radius of the cone.
}
\description{
Calculates the radius of the cone based on the input height.
}
\examples{
cone_radius(5)
cone_radius(-1)
}
\author{
Sofia Daza
}
