% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapply_function.R
\name{cone_radius_sapply}
\alias{cone_radius_sapply}
\title{Cone Radius Function using sapply}
\usage{
cone_radius_sapply(x_values)
}
\arguments{
\item{x_values}{Numeric vector of heights.}
}
\value{
A numeric vector of radii.
}
\description{
Functions for calculating the radius of the cone using different approaches.
}
\examples{
cone_radius_for(seq(0, 10, length.out = 100))
}
\author{
Sofia Daza
}
