% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivative.R
\name{derivative}
\alias{derivative}
\title{Numerical Derivative of Cone Radius}
\usage{
derivative(x, epsilon = 1e-05)
}
\arguments{
\item{x}{A numeric value representing the height.}

\item{epsilon}{A small numeric value used for numerical differentiation. Default is 1e-5.}
}
\value{
A numeric value representing the approximate derivative.
}
\description{
Approximates the derivative of the `cone_radius_ifelse` function at a given height.
}
\examples{
derivative(5)
derivative(8)
}
