% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_functions.R
\name{lookup_weather_data}
\alias{lookup_weather_data}
\title{Lookup Weather Data}
\usage{
lookup_weather_data(city)
}
\arguments{
\item{city}{A string representing the name of the city.}
}
\value{
A list containing the weather data or `NULL` if the request fails.
}
\description{
Retrieves weather forecast data for a given city from the OpenWeatherMap API.
}
\author{
Sofia Daza
}
