% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulation}
\alias{simulation}
\title{Party Simulation}
\usage{
simulation(num_guests)
}
\arguments{
\item{num_guests}{Number of guests attending the party.}
}
\value{
Simulated total cones and guest information.
}
\description{
Functions to simulate party resources based on weather forecasts.
}
\author{
Sofia Daza
}
