% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_area.R
\name{surface_area}
\alias{surface_area}
\title{Surface Area of the Cone}
\usage{
surface_area()
}
\value{
A numeric value representing the surface area of the cone.
}
\description{
Calculates the surface area of the cone using numerical integration.
}
\examples{
surface_area()
}
