% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iu.R
\docType{data}
\name{ui}
\alias{ui}
\title{UI for the Shiny App}
\format{
An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 4.
}
\usage{
ui
}
\description{
User interface definition for the Ice Cream Party Simulation app.
}
\author{
Sofia Daza
}
\keyword{datasets}
