% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume_surface_area.R
\name{volume}
\alias{volume}
\title{Volume and Surface Area Calculations}
\usage{
volume()
}
\value{
The calculated volume or surface area.
}
\description{
Functions to calculate the volume and surface area of a cone.
}
\examples{
volume()
}
\author{
Sofia Daza
}
