% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cone_radius_functions.R
\name{cone_radius}
\alias{cone_radius}
\title{Cone Radius Non Vectorized Function}
\usage{
cone_radius(x)
}
\arguments{
\item{x_values}{height}
}
\value{
A radius
}
\description{
Functions for calculating the radius of the cone using different approaches.
}
\examples{
2
}
\author{
Sofia Daza
}
