% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize_function.R
\name{cone_radius_vectorize}
\alias{cone_radius_vectorize}
\title{Cone Radius Function Vectorize}
\usage{
cone_radius_vectorize(x)
}
\arguments{
\item{x_values}{Numeric vector of heights.}
}
\value{
A numeric vector of radii.
}
\description{
Functions for calculating the radius of the cone using different approaches.
}
\examples{
cone_radius_for(seq(0, 10, length.out = 100))
}
\author{
Sofia Daza
#' @export
}
