% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surface_area.R
\name{derivative}
\alias{derivative}
\title{Surface Area Calculations}
\usage{
derivative(x, epsilon = 1e-05)
}
\value{
The calculated volume or surface area.
}
\description{
Functions to calculate the volume and surface area of a cone.
}
\examples{
volume()
}
\author{
Sofia Daza
}
