% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weather_functions.R
\name{lookup_weather_data}
\alias{lookup_weather_data}
\title{Weather Data Retrieval}
\usage{
lookup_weather_data(city)
}
\arguments{
\item{city}{A character string for the city name.}
}
\value{
A list containing current and forecast weather data.
}
\description{
Functions to retrieve and process weather data using the OpenWeather API.
}
\examples{
lookup_weather_data("Lausanne")
}
\author{
Sofia Daza
}
