% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulation}
\alias{simulation}
\title{Simulation Function}
\usage{
simulation(n, prob)
}
\arguments{
\item{n}{Number of trials.}

\item{prob}{Probability of success.}
}
\value{
A vector of simulated outcomes.
}
\description{
Simulates some process using `rbinom`.
}
\examples{
simulation(100, 0.5)
}
